+++ src/DataProvider/Select/DataProvider.php
@@ -8,20 +8,21 @@ use ddGetDocuments\Input;
 
 class DataProvider extends \ddGetDocuments\DataProvider\DataProvider
 {
-	public $defaultParams = array(
+	public $defaultParams = [
 		'ids' => null
-	);
+	];
 	
 	/**
 	 * getDataFromSource
+	 * @version 1.0.3 (2016-12-21)
 	 * 
-	 * @param Input $input
+	 * @param $input {ddGetDocuments\Input}
 	 * 
-	 * @return Output
+	 * @return {ddGetDocuments\DataProvider\Output}
 	 */
 	protected function getDataFromSource(Input $input){
 		global $modx;
-		$output = new Output(array(), 0);
+		$output = new Output([], 0);
 		
 		$ids = $this->defaultParams['ids'];
 		
@@ -50,7 +51,7 @@ class DataProvider extends \ddGetDocuments\DataProvider\DataProvider
 		}
 		
 		//By default, the required data is just fetched from the site_content table
-		$fromQuery = "{$this->siteContentTableName}";
+		$fromQuery = $this->siteContentTableName;
 		$filterQuery = '';
 		
 		//If a filter is set, it is needed to check which TVs are used in the filter query
@@ -60,55 +61,71 @@ class DataProvider extends \ddGetDocuments\DataProvider\DataProvider
 			//If there are some TV names in the filter query, make a temp table from which the required data will be fetched
 			if(!empty($usedFields['tvs'])){
 				//complete from query
-				$fromQuery = "(".$this->buildTVsSubQuery($usedFields['tvs']).")";
+				$fromQuery = '('.$this->buildTVsSubQuery($usedFields['tvs']).')';
 			}
 			
-			$filterQuery = "$filter";
+			$filterQuery = $filter;
 			$filterQuery = str_replace($fieldDelimiter, '`', $filterQuery);
 		}
 		
 		$orderByQuery = '';
 		
 		if(!empty($orderBy)){
-			$orderByQuery = "ORDER BY $orderBy";
+			$orderByQuery = 'ORDER BY '.$orderBy;
+		//Order by selected IDs sequence
+		}elseif(!empty($ids)){
+			$orderByQuery = 'ORDER BY FIELD (`documents`.`id`,'.$ids.')';
 		}
 		
 		$limitQuery = '';
 		
-		if(!empty($offset) && !empty($total)){
-			$limitQuery = "LIMIT $offset,$total";
-		}elseif(empty($offset) && !empty($total)){
-			$limitQuery = "LIMIT $total";
-		}elseif(!empty($offset) && empty($total)){
-			$limitQuery = "LIMIT $offset,".PHP_INT_MAX;
+		if(
+			!empty($offset) &&
+			!empty($total)
+		){
+			$limitQuery = 'LIMIT '.$offset.','.$total;
+		}elseif(
+			empty($offset) &&
+			!empty($total)
+		){
+			$limitQuery = 'LIMIT '.$total;
+		}elseif(
+			!empty($offset) &&
+			empty($total)
+		){
+			$limitQuery = 'LIMIT '.$offset.','.PHP_INT_MAX;
 		}
 		
 		$idsWhereQuery = '';
 		if(!empty($ids)){
-			$idsWhereQuery = "`documents`.`id` IN ($ids)";
+			$idsWhereQuery = '`documents`.`id` IN ('.$ids.')';
 		}
 		
 		$whereQuery = '';
-		if(!empty($idsWhereQuery) || !empty($filterQuery)){
-			$whereQuery = "WHERE ";
+		if(
+			!empty($idsWhereQuery) ||
+			!empty($filterQuery)
+		){
+			$whereQuery = 'WHERE ';
 			if(!empty($idsWhereQuery)){
 				$whereQuery .= $idsWhereQuery;
 				
 				if(!empty($filterQuery)){
-					$whereQuery .= " AND $filterQuery";
+					$whereQuery .= ' AND '.$filterQuery;
 				}
 			}else{
 				$whereQuery .= $filterQuery;
 			}
 		}
 		
-		$data = $modx->db->makeArray($modx->db->query("
-			SELECT SQL_CALC_FOUND_ROWS `documents`.`id` FROM $fromQuery AS `documents`
-			$whereQuery $orderByQuery $limitQuery
-		"));
+		$data = $modx->db->makeArray($modx->db->query('
+			SELECT
+				SQL_CALC_FOUND_ROWS `documents`.`id`
+			FROM '.$fromQuery.' AS `documents`
+			'.$whereQuery.' '.$orderByQuery.' '.$limitQuery.'
+		'));
 		
-		$totalFoundArray = $modx->db->makeArray($modx->db->query("SELECT FOUND_ROWS() as `totalFound`"));
-		$totalFound = $totalFoundArray[0]['totalFound'];
+		$totalFound = $modx->db->getValue('SELECT FOUND_ROWS()');
 		
 		if(is_array($data)){
 			$output = new Output($data, $totalFound);